requirements.txt
python-dotenv
bitcoinrpc
asyncio
venv